Time: 28 ms (44.71%) | Memory: 55.3 MB (38.79%) - LeetSync

+++ 1621-number-of-subsequences-that-satisfy-the-given-sum-condition/number-of-subsequences-that-satisfy-the-given-sum-condition.java
@@ -24,6 +24,6 @@ class Solution {
             }
         }
 
-        return (int)tot;
+        return tot;
     }
 }
